Moving container-level components into the container folder.
