--- app/components/Dashboard.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default class Dashboard extends React.Component {
-  render() {
-    return (
-      <div>Welcome to the Dashboard</div>
-    );
-  }
-}
