+++ app/constants/index.js
@@ -1,6 +1,7 @@
 export const CREATE_TOPIC_REQUEST = 'CREATE_TOPIC_REQUEST';
 export const CREATE_TOPIC_FAILURE = 'CREATE_TOPIC_FAILURE';
 export const CREATE_TOPIC_SUCCESS = 'CREATE_TOPIC_SUCCESS';
+export const CREATE_TOPIC_DUPLICATE = 'CREATE_TOPIC_DUPLICATE';
 export const INCREMENT_COUNT = 'INCREMENT_COUNT';
 export const DECREMENT_COUNT = 'DECREMENT_COUNT';
 export const DESTROY_TOPIC = 'DESTROY_TOPIC';
