+++ app/containers/About.jsx
@@ -4,6 +4,11 @@ import styles from 'scss/components/_about';
 
 const cx = classNames.bind(styles);
 
+/*
+ * Note: This is kept as a container-level component, 
+ *  i.e. We should keep this as the container that does the data-fetching 
+ *  and dispatching of actions if you decide to have any sub-components.
+ */
 export default class About extends React.Component {
   render() {
     return (
@@ -13,4 +18,4 @@ export default class About extends React.Component {
       </div>
     );
   }
-}
+};
