+++ app/containers/Logout.jsx
@@ -1,6 +1,11 @@
 import React, { Component } from 'react';
 import styles from 'scss/components/_logout';
 
+/*
+ * Note: This is kept as a container-level component, 
+ *  i.e. We should keep this as the container that does the data-fetching 
+ *  and dispatching of actions if you decide to have any sub-components.
+ */
 export default class Logout extends Component {
   render() {
     return (
@@ -9,4 +14,4 @@ export default class Logout extends Component {
       </div>
     );
   }
-}
+};
