+++ app/containers/Vote.jsx
